genai-backend/pyproject.toml
[project]
name = "genai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bson>=0.5.10",
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "email-validator>=2.0.0",
    "fastapi>=0.117.1",
    "fitz>=0.0.1.dev2",
    "google-genai>=1.38.0",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.15.1",
    "pypdf2>=3.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pytesseract>=0.3.13",
    "reportlab>=4.0.0",
    "uvicorn>=0.36.0",
]
